requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-dateutil>=2.9.0
pydantic>=2.0.0
psycopg2-binary>=2.9.9
